+++ backtest/__init__.py
@@ -1,0 +1,1 @@
+"""Backtest MVP package for StockAnalyzerPro."""